feature: add qr code scanner

+++ lib/history/application/bloc/history_bloc.dart
@@ -7,20 +7,18 @@ import 'package:fast_shopping/models/fs_product.dart';
 import 'package:fast_shopping/models/fs_user.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:injectable/injectable.dart';
-import 'package:intl/intl.dart';
 
-part 'basket_bloc.freezed.dart';
-part 'basket_event.dart';
-part 'basket_state.dart';
+part 'history_bloc.freezed.dart';
+part 'history_event.dart';
+part 'history_state.dart';
 
 @injectable
-class BasketBloc extends Bloc<BasketEvent, BasketState> {
-  BasketBloc({
+class HistoryBloc extends Bloc<HistoryEvent, HistoryState> {
+  HistoryBloc({
     required this.productRepository,
-  }) : super(const BasketState()) {
+  }) : super(const HistoryState()) {
     on<GetUser>(_getUser);
     on<GetProducts>(_getProducts);
-    on<PayProduct>(_payProduct);
     on<DeleteProduct>(_deleteProduct);
     on<EmitUser>(_emitUser);
     on<EmitProduct>(_emitProduct);
@@ -31,7 +29,7 @@ class BasketBloc extends Bloc<BasketEvent, BasketState> {
 
   late final StreamSubscription<List<FSOrder?>> _productStream;
 
-  Future<void> _getUser(GetUser event, Emitter<BasketState> emit) async {
+  Future<void> _getUser(GetUser event, Emitter<HistoryState> emit) async {
     final user = await productRepository.getUser();
 
     productRepository.getUserStream(user.uid).listen((event) {
@@ -40,7 +38,7 @@ class BasketBloc extends Bloc<BasketEvent, BasketState> {
   }
 
   Future<void> _getProducts(
-      GetProducts event, Emitter<BasketState> emit) async {
+      GetProducts event, Emitter<HistoryState> emit) async {
     _productStream = productRepository.getOrdersStream().listen((event) {
       add(EmitProduct(event));
     });
@@ -53,78 +51,40 @@ class BasketBloc extends Bloc<BasketEvent, BasketState> {
     _productStream.cancel();
   }
 
-  Future<void> _payProduct(PayProduct event, Emitter<BasketState> emit) async {
-    final f = DateFormat('yyyy-MM-dd hh:mm');
-    List<String> productsHistory = [];
-    List<String> products = [];
-
-    if (state.user == null) {
-      final FSUser user = await productRepository.getUser();
-
-      productsHistory.addAll(user.productsHistory);
-      products.addAll(user.products);
-      FSOrder oldOrder = state.orders
-          .firstWhere((element) => element.orderId == event.product);
-
-      final order = await productRepository.buyProduct(FSOrder(
-          orderId: '',
-          productId: oldOrder.productId,
-          date: f.format(DateTime.now())));
-
-      productsHistory.add(order.orderId);
-
-      products.remove(oldOrder.orderId);
-      await productRepository.deleteOrder(order.orderId);
-      final newUser =
-          user.copyWith(productsHistory: productsHistory, products: products);
-      final updateUser = await productRepository.updateUser(newUser);
-    } else {
-      final FSUser user = await productRepository.getUser();
-
-      productsHistory.addAll(user.productsHistory);
-      products.addAll(user.products);
-      FSOrder oldOrder = state.orders
-          .firstWhere((element) => element.orderId == event.product);
-
-      final order = await productRepository.buyProduct(FSOrder(
-          orderId: '',
-          productId: oldOrder.productId,
-          date: f.format(DateTime.now())));
-      productsHistory.add(order.orderId);
-
-      products.remove(oldOrder.orderId);
-      await productRepository.deleteOrder(oldOrder.orderId);
-      final newUser =
-          user.copyWith(productsHistory: productsHistory, products: products);
-      final updateUser = await productRepository.updateUser(newUser);
-    }
-  }
-
   Future<void> _deleteProduct(
     DeleteProduct event,
-    Emitter<BasketState> emit,
+    Emitter<HistoryState> emit,
   ) async {
     if (state.user != null) {
-      List<String> newBasket = state.user!.products;
-      newBasket.remove(event.id);
-      await productRepository
-          .updateUser(state.user!.copyWith(products: newBasket));
-      await productRepository.deleteOrder(event.id);
+      List<String> oldBasket = state.user!.productsHistory;
+
+      for (String product in state.user!.productsHistory) {
+        if (product == event.id) {
+          oldBasket.remove(product);
+          break;
+        }
+
+        await productRepository
+            .updateUser(state.user!.copyWith(productsHistory: oldBasket));
+        await productRepository.deleteOrder(event.id);
+        add(EmitUser(state.user!));
+      }
     } else {
       final user = await productRepository.getUser();
+
       user.products.remove(event.id);
       await productRepository.updateUser(user);
-      await productRepository.deleteOrder(event.id);
+      add(EmitUser(user));
     }
   }
 
-  Future<void> _emitUser(EmitUser event, Emitter<BasketState> emit) async {
+  Future<void> _emitUser(EmitUser event, Emitter<HistoryState> emit) async {
     List<FSProduct> sortList = [];
     List<FSOrder> sortListOrders = [];
 
     final user = await productRepository.getUser();
 
-    for (String basket in user.products) {
+    for (String basket in user.productsHistory) {
       for (FSOrder product in await productRepository.getOrdersStream().first) {
         if (basket == product.orderId) {
           sortList.add(await productRepository.getProduct(product.productId));
@@ -140,13 +100,14 @@ class BasketBloc extends Bloc<BasketEvent, BasketState> {
   }
 
   FutureOr<void> _emitProduct(
-      EmitProduct event, Emitter<BasketState> emit) async {
+      EmitProduct event, Emitter<HistoryState> emit) async {
     List<FSProduct> sortList = [];
     List<FSOrder> sortListOrders = [];
+
     if (state.user == null) {
       final user = await productRepository.getUser();
 
-      for (String basket in user.products) {
+      for (String basket in user.productsHistory) {
         for (FSOrder product in event.product) {
           if (basket == product.orderId) {
             sortList.add(await productRepository.getProduct(product.productId));
@@ -155,7 +116,7 @@ class BasketBloc extends Bloc<BasketEvent, BasketState> {
         }
       }
     } else {
-      for (String basket in state.user!.products) {
+      for (String basket in state.user!.productsHistory) {
         for (FSOrder product in event.product) {
           if (basket == product.orderId) {
             sortList.add(await productRepository.getProduct(product.productId));
@@ -164,7 +125,6 @@ class BasketBloc extends Bloc<BasketEvent, BasketState> {
         }
       }
     }
-
     emit(state.copyWith(
       products: sortList,
       orders: sortListOrders,
